+++ src/app/project/list/project.list.component.ts
@@ -1,7 +1,8 @@
 import 'rxjs/Rx';
 import { BehaviorSubject } from 'rxjs/BehaviorSubject';
+import { Observable } from 'rxjs/Observable';
 import { Component, Input, OnChanges, ChangeDetectionStrategy } from '@angular/core';
-import { ProjectService, ProjectModel } from '../../shared/project/index';
+import { ProjectService, ProjectModel } from '../../shared/project';
 import { Http, Response, Headers, RequestOptions } from '@angular/http';
 import { UserService } from '../../shared/user/user.service';
 
@@ -15,17 +16,24 @@ import { UserService } from '../../shared/user/user.service';
 export class ProjectListComponent implements OnChanges {
 
   @Input() leaderId;
-  @Input() maxCount = 100;
+  @Input() pageSize = 5;
 
-  private projects: BehaviorSubject<any> = new BehaviorSubject([{title: 'Loading...'}]);
+  public projects: BehaviorSubject<any> = new BehaviorSubject([{title: 'Loading...'}]);
+  public itemsPage = {
+    docs: this.projects,
+    limit: this.pageSize,
+    page: 1,
+    pages: 0,
+    total: 0
+  };
 
   isAddingTaskMode = false;
 
   ngOnChanges(changes) {
     if (changes.leaderId && changes.leaderId.currentValue ) {
-      this.requestProjects(changes.leaderId.currentValue);
-    } else if (changes.maxCount && changes.maxCount.currentValue) {
-      this.requestProjects(null, changes.maxCount.currentValue);
+      this.requestProjects();
+    } else if (changes.pageSize && changes.pageSize.currentValue) {
+      this.requestProjects();
     }
   }
 
@@ -35,12 +43,23 @@ export class ProjectListComponent implements OnChanges {
     private http: Http
   ) {}
 
+  pageChanged(pageNumber) {
+    this.itemsPage.page = pageNumber;
+    this.requestProjects();
+  }
+
   // WIP
-  requestProjects(leaderId = '', maxCount = 100) {
-    const proxySub = this.projectService.getProjects('', leaderId, maxCount).subscribe(projects => {
-      this.projects.next(projects);
-      proxySub.unsubscribe();
-    });
+  requestProjects() {
+    const proxySub = this.projectService.getProjectsPage(null, this.leaderId, this.itemsPage.page, this.pageSize)
+      .subscribe(responsePage => {
+        // console.log('Next, responsePage:', responsePage);
+        this.itemsPage.docs.next(responsePage['docs']);
+        this.itemsPage.limit = responsePage['limit'];
+        this.itemsPage.page = responsePage['page'];
+        this.itemsPage.pages = responsePage['pages'];
+        this.itemsPage.total = responsePage['total'];
+        proxySub.unsubscribe();
+      });
   }
 
   deleteProject(projectToRemove: ProjectModel) {